feat(cli): ship version management in every build; only the net downloader is full-build-only
