+++ cli/wago/version_common_test.go
@@ -1,12 +1,6 @@
-//go:build !wago_lean
-
 package main
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"path/filepath"
 	"testing"
@@ -51,46 +45,6 @@ func TestVersionManagerState(t *testing.T) {
 	}
 	if activeVersion(d) != "0.5.0" {
 		t.Fatalf("activeVersion = %q, want 0.5.0", activeVersion(d))
-	}
-}
-
-func TestDownloadBinaryChecksum(t *testing.T) {
-	payload := []byte("fake wago binary bytes")
-	sum := sha256.Sum256(payload)
-	hexsum := hex.EncodeToString(sum[:])
-
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch {
-		case r.URL.Path == "/0.9.0/wago-linux-amd64":
-			w.Write(payload)
-		case r.URL.Path == "/0.9.0/wago-linux-amd64.sha256":
-			w.Write([]byte(hexsum + "  wago-linux-amd64\n"))
-		case r.URL.Path == "/bad/wago-linux-amd64":
-			w.Write(payload)
-		case r.URL.Path == "/bad/wago-linux-amd64.sha256":
-			w.Write([]byte("deadbeef  wago-linux-amd64\n"))
-		default:
-			http.NotFound(w, r)
-		}
-	}))
-	defer srv.Close()
-
-	dest := filepath.Join(t.TempDir(), "wago")
-	if err := downloadBinary(srv.URL, "0.9.0", dest); err != nil {
-		t.Fatalf("downloadBinary: %v", err)
-	}
-	got, err := os.ReadFile(dest)
-	if err != nil || string(got) != string(payload) {
-		t.Fatalf("downloaded content mismatch: %v", err)
-	}
-
-	// A checksum mismatch must fail and write nothing.
-	badDest := filepath.Join(t.TempDir(), "wago")
-	if err := downloadBinary(srv.URL, "bad", badDest); err == nil {
-		t.Fatal("expected checksum mismatch error")
-	}
-	if _, err := os.Stat(badDest); !os.IsNotExist(err) {
-		t.Fatal("checksum mismatch must not write the destination file")
 	}
 }
